BREAKING: This version does not send the display to sleep, only delays then restarts. The server now only supports random images for each day.
I had a ton of issues making the display reliably come out of sleep and also making the display go to sleep to not consume current. I ended up just making the frame USB powered to go on a desk and just sit idle while waiting to show next image.